moved over to reactRouter
setup the router in the app.js component, inorder to get the props to
properly work, had to create a wrapper component, typingPage for the
whole typing page. Also had to rename some levels in order to get rid of
symbols such as # in the title which interfere in the routing. Currenlty
the /browse page lists all the levels, and /type/:level is how to get a
a certain level